+++ sdk/src/main/java/ly/count/android/sdk/ModuleContent.java
@@ -12,11 +12,16 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import org.jetbrains.annotations.NotNull;
 import org.json.JSONObject;
 
 public class ModuleContent extends ModuleBase {
     private final ImmediateRequestGenerator iRGenerator;
+    private final ExecutorService refreshExecutor; // to not block main thread during refresh
+    Future<?> refreshContentZoneInternalFuture; // for test access
     Content contentInterface;
     CountlyTimer countlyTimer;
     private boolean shouldFetchContents = false;
@@ -25,12 +30,12 @@ public class ModuleContent extends ModuleBase {
     private final ContentCallback globalContentCallback;
     private int waitForDelay = 0;
     int CONTENT_START_DELAY_MS = 4000; // 4 seconds
-    int REFRESH_CONTENT_ZONE_DELAY_MS = 2500; // 2.5 seconds
 
     ModuleContent(@NonNull Countly cly, @NonNull CountlyConfig config) {
         super(cly, config);
         L.v("[ModuleContent] Initialising, zoneTimerInterval: [" + config.content.zoneTimerInterval + "], globalContentCallback: [" + config.content.globalContentCallback + "]");
         iRGenerator = config.immediateRequestGenerator;
+        refreshExecutor = Executors.newSingleThreadExecutor();
 
         contentInterface = new Content();
         countlyTimer = new CountlyTimer();
@@ -295,6 +300,7 @@ public class ModuleContent extends ModuleBase {
         contentInterface = null;
         countlyTimer.stopTimer(L);
         countlyTimer = null;
+        refreshExecutor.shutdown();
     }
 
     @Override
@@ -329,13 +335,13 @@ public class ModuleContent extends ModuleBase {
             return;
         }
 
-        if (!shouldFetchContents) {
-            exitContentZoneInternal();
-        }
+        exitContentZoneInternal();
 
-        _cly.moduleRequestQueue.attemptToSendStoredRequestsInternal();
-
-        enterContentZoneInternal(null, REFRESH_CONTENT_ZONE_DELAY_MS);
+        refreshContentZoneInternalFuture = refreshExecutor.submit(() -> {
+            _cly.moduleRequestQueue.attemptToSendStoredRequestsInternal(true);
+            L.d("[ModuleContent] refreshContentZone, RQ flush done, re-entering content zone");
+            enterContentZoneInternal(null, 0);
+        });
     }
 
     public class Content {