+++ sdk/src/main/java/ly/count/android/sdk/RequestQueueProvider.java
@@ -52,6 +52,8 @@ interface RequestQueueProvider {
 
     void tick();
 
+    void tick(boolean forceFlushRQ);
+
     ConnectionProcessor createConnectionProcessor();
 
     String prepareRemoteConfigRequestLegacy(@Nullable String keysInclude, @Nullable String keysExclude, @NonNull String preparedMetrics);