+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX.XX.XX
+* Improved Content refresh mechanics.
+
 ## 25.4.9
 * Added a new config option `disableViewRestartForManualRecording()` to disable auto close/restart behavior of manual views on app background/foreground actions.